requirements.txt
streamlit
python-dotenv
openai
reportlab
python-docx
markdown2
beautifulsoup4
requests
bs4